config/requirements.txt
python-dotenv>=1.0.1
psycopg2-binary>=2.9.11
hdbscan>=0.8.41
numpy>=2.0.2
scikit-learn>=1.6.1
scipy>=1.13.1
PyYAML>=6.0.3
openai>=1.68.2
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic-settings>=2.4.0
sentence-transformers>=3.0.0
spacy>=3.7.0
requests>=2.32.0
google-cloud-secret-manager>=2.20.0
google-cloud-pubsub>=2.21.0
